create utils.distributions
